requirements.txt
beautifulsoup4==4.12.2
openai==1.3.5
python-dotenv==1.0.0
